% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backward_procedure_cov.R
\name{backward_procedure_cov}
\alias{backward_procedure_cov}
\title{Estimates the Backward probabilities}
\usage{
backward_procedure_cov(Data, X_cov, n, r, k, OHC, kOHC, Pi, A, Phi, Phi_cov, P,
  Theta, Q, spec, ...)
}
\arguments{
\item{Data}{is a single univariate data sequence}

\item{X_cov}{contains the sequence(s) of visible level covariates (in a matrix form)}
}
\value{
1. scaled Backward probabilities, 2. scaling constants
}
\description{
Estimates the Backward probabilities
}
