% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backward_procedure_NOcov.R
\name{backward_procedure_NOcov}
\alias{backward_procedure_NOcov}
\title{Estimates the Backward probabilities (without visible level covariates)}
\usage{
backward_procedure_NOcov(Data, n, r, k, OHC, kOHC, Pi, A, Phi, P, Theta, Q,
  spec, ...)
}
\arguments{
\item{Data}{is a single univariate data sequence}
}
\value{
1. scaled backward probabilities, 2. scaling constants
}
\description{
Estimates the Backward probabilities (without visible level covariates)
}
