% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epsilon_cov.R
\name{epsilon_cov}
\alias{epsilon_cov}
\title{Estimates the          probabilities}
\usage{
epsilon_cov(Data, X_cov, n, r, k, OHC, kOHC, Pi, A, SAlpha, SAlog, SBeta, SBlog,
  LLAlpha, Phi, Phi_cov, Theta, P, Q, spec)
}
\arguments{
\item{Data}{is a single univariate data sequence}

\item{X_cov}{contains the sequence(s) of visible level covariates (in a matrix form)}
}
\value{
Epsilon probabilities
}
\description{
Estimates the          probabilities
}
