% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epsilon_NOcov.R
\name{epsilon_NOcov}
\alias{epsilon_NOcov}
\title{Estimates the}
\usage{
epsilon_NOcov(Data, n, r, k, OHC, kOHC, Pi, A, SAlpha, SAlog, SBeta, SBlog,
  LLAlpha, Phi, Theta, P, Q, spec)
}
\arguments{
\item{Data}{is a single univariate data sequence}
}
\value{
epsilon probabilities
}
\description{
Estimates the
}
