% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forward_procedure_NOcov.R
\name{forward_procedure_NOcov}
\alias{forward_procedure_NOcov}
\title{Estimates the Forward probabilities (without visible level covariates)}
\usage{
forward_procedure_NOcov(Data, n, r, k, OHC, kOHC, Pi, A, Phi, P, Theta, Q, spec,
  ...)
}
\arguments{
\item{Data}{is a single univariate data sequence}
}
\value{
1. scaled forward probabilities, 2. scaling constants and 3.Incomplete data log-Likelihood
}
\description{
Estimates the Forward probabilities (without visible level covariates)
}
