% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMTD_model.R
\name{HMTD_model}
\alias{HMTD_model}
\title{HMTD model estimation}
\usage{
HMTD_model(Data, X_cov, Phi_cov, phi_cov_places, k, OHC, P, Q,
  procedure = c("SEM_v2"), it, envir, Adiag, covmin = NA, covmax = NA,
  initialVECTPAR = NULL, picov = NA, ...)
}
\arguments{
\item{Data}{is a matrix of data sequences (in rows)}

\item{X_cov}{contains the sequence(s) of visible level covariates (in a matrix form)}
}
\value{
1. scaled forward probabilities, 2. scaling constants and 3. incomplete log-Likelihood
}
\description{
HMTD model estimation
}
