% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmtd_sigma_arch_nocov.R
\name{hmtd_sigma_arch_nocov}
\alias{hmtd_sigma_arch_nocov}
\title{Estimates the std of the components}
\usage{
hmtd_sigma_arch_nocov(Theta, Phi, p, Data)
}
\arguments{
\item{Theta}{Theta}

\item{Phi}{Phi}
}
\value{
std
}
\description{
Estimates the std of the components
}
