% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/la_nouvelle_approche_max_AVEC_CHANGEMENTS_PRECEDENTS_maxiterations.R
\name{la_nouvelle_approche_max_AVEC_CHANGEMENTS_PRECEDENTS_maxiterations}
\alias{la_nouvelle_approche_max_AVEC_CHANGEMENTS_PRECEDENTS_maxiterations}
\title{Maximization procedure "S"}
\usage{
la_nouvelle_approche_max_AVEC_CHANGEMENTS_PRECEDENTS_maxiterations(fct, VECTPAR,
  target, maxdelta, mindelta, it, envir...)
}
\arguments{
\item{fct}{function to be maximized (logLikelihood)}

\item{VECTPAR}{vector of visiible parameters Phi and Theta}
}
\value{
logL, parameters, iterations and number of function calls
}
\description{
Maximization procedure "S"
}
