% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/nouvelle_appr_consecutive_modification_v2_maxiterations.R
\name{nouvelle_appr_consecutive_modification_v2_maxiterations}
\alias{nouvelle_appr_consecutive_modification_v2_maxiterations}
\title{Maximization procedure SEM}
\usage{
nouvelle_appr_consecutive_modification_v2_maxiterations(fct, VECTPAR, lengthPhi,
  lengthTheta, k, target, maxdelta, mindelta, it, envir...)
}
\arguments{
\item{fct}{function to be maximized (logLikelihood)}

\item{VECTPAR}{vector of visiible parameters Phi and Theta}
}
\value{
logL, parameters, iterations and number of function calls
}
\description{
Maximization procedure SEM
}
