% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEM_v2.R
\name{SEM_v2}
\alias{SEM_v2}
\title{Maximization procedure "SEM_v2"}
\usage{
SEM_v2(fct, VECTPAR, lengthPhi, lengthTheta, k, target = NA, it, envir,
  lower = VECT_limits_min, upper = VECT_limits_max, mutheta, sigtheta,
  muphi, sigphi, initialVECTPAR = NULL)
}
\arguments{
\item{fct}{function to be maximized (logLikelihood)}

\item{VECTPAR}{vector of visiible parameters Phi and Theta}
}
\value{
logL, parameters, iterations and number of function calls
}
\description{
Maximization procedure "SEM_v2"
}
