% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEMP_v2.R
\name{SEMP_v2}
\alias{SEMP_v2}
\title{Maximization procedure "SEMP_v2"}
\usage{
SEMP_v2(fct, VECTPAR, lengthPhi, lengthTheta, k, target = NA, it, envir,
  lower = VECT_limits_min, upper = VECT_limits_max)
}
\arguments{
\item{fct}{function to be maximized (logLikelihood)}

\item{VECTPAR}{vector of visiible parameters Phi and Theta}
}
\value{
logL, parameters, iterations and number of function calls
}
\description{
Maximization procedure "SEMP_v2"
}
